Update README.md
The code successfully helps automate manual tasks and thus reduce the time it takes to complete reporting from approximately 2 hours per today to only 20 minutes.

The initial part of the code automatically transfers the raw data files received in Adobe, Demyst outlook inbox into it's respective file path which is created dynamically parsing raw file names and it's received date/time. The code also runs a Microsoft VBA which via a function called 'run_downgrader_tool'.

Thereafter the code connects to an Oracle database without hardcoding the log in credentials and executes 2 SQL queries. 

The output of first SQL query is used in the second SQL query and the final output is structured using the Pandas library and stored in an excel file which is saved in the shared file path.